Add other pages to landing
